requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
dash==4.0.0
Flask==3.1.3
fonttools==4.61.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.17.0
nest-asyncio==1.6.0
numpy==2.2.6
openseespy==3.7.1.2
openseespylinux==3.7.1.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
scipy==1.15.3
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
torch==2.10.0
scikit-learn==1.7.2
